move CompiledInstruction to message module
