+++ sdk/program/src/instruction.rs
@@ -1,59 +1,12 @@
 #[cfg(feature = "frozen-abi")]
 use solana_frozen_abi_macro::AbiExample;
-pub use solana_instruction::{
-    error::InstructionError, AccountMeta, Instruction, ProcessedSiblingInstruction,
-    TRANSACTION_LEVEL_STACK_HEIGHT,
+pub use {
+    crate::message::compiled_instruction::CompiledInstruction,
+    solana_instruction::{
+        error::InstructionError, AccountMeta, Instruction, ProcessedSiblingInstruction,
+        TRANSACTION_LEVEL_STACK_HEIGHT,
+    },
 };
-use {
-    bincode::serialize, serde::Serialize, solana_pubkey::Pubkey, solana_sanitize::Sanitize,
-    solana_short_vec as short_vec,
-};
-
-/// A compact encoding of an instruction.
-///
-/// A `CompiledInstruction` is a component of a multi-instruction [`Message`],
-/// which is the core of a Solana transaction. It is created during the
-/// construction of `Message`. Most users will not interact with it directly.
-///
-/// [`Message`]: crate::message::Message
-#[cfg_attr(feature = "frozen-abi", derive(AbiExample))]
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct CompiledInstruction {
-    /// Index into the transaction keys array indicating the program account that executes this instruction.
-    pub program_id_index: u8,
-    /// Ordered indices into the transaction keys array indicating which accounts to pass to the program.
-    #[serde(with = "short_vec")]
-    pub accounts: Vec<u8>,
-    /// The program input data.
-    #[serde(with = "short_vec")]
-    pub data: Vec<u8>,
-}
-
-impl Sanitize for CompiledInstruction {}
-
-impl CompiledInstruction {
-    pub fn new<T: Serialize>(program_ids_index: u8, data: &T, accounts: Vec<u8>) -> Self {
-        let data = serialize(data).unwrap();
-        Self {
-            program_id_index: program_ids_index,
-            accounts,
-            data,
-        }
-    }
-
-    pub fn new_from_raw_parts(program_id_index: u8, data: Vec<u8>, accounts: Vec<u8>) -> Self {
-        Self {
-            program_id_index,
-            accounts,
-            data,
-        }
-    }
-
-    pub fn program_id<'a>(&self, program_ids: &'a [Pubkey]) -> &'a Pubkey {
-        &program_ids[self.program_id_index as usize]
-    }
-}
 
 /// Returns a sibling instruction from the processed sibling instruction list.
 ///
